requirements.txt
# Core deep learning framework
tensorflow==2.15.0

# Data manipulation and numerical computation
numpy>=1.23.5
pandas>=1.5.0

# Image processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0

# Machine learning utilities
scikit-learn>=1.1.0
tensorflow-datasets>=4.8.0

# File handling and downloading
wget>=3.2

# Development and testing
jupyter>=1.0.0
notebook>=6.4.0

# Optional: For additional functionality
h5py>=3.7.0
absl-py>=1.0.0